Issue #3165 - rearrange packages for javax-websocket-server tests
