+++ jetty-websocket/javax-websocket-server/src/test/java/org/eclipse/jetty/websocket/javax/server/examples/MyAuthedSocket.java
@@ -16,19 +16,18 @@
 //  ========================================================================
 //
 
-package examples;
+package org.eclipse.jetty.websocket.javax.server.examples;
 
-import javax.servlet.http.HttpSession;
-import javax.websocket.HandshakeResponse;
-import javax.websocket.server.HandshakeRequest;
-import javax.websocket.server.ServerEndpointConfig;
+import javax.websocket.OnMessage;
+import javax.websocket.server.ServerEndpoint;
 
-public class GetHttpSessionConfigurator extends ServerEndpointConfig.Configurator
+@ServerEndpoint(value = "/secured/socket", configurator = MyAuthedConfigurator.class)
+public class MyAuthedSocket
 {
-    @Override
-    public void modifyHandshake(ServerEndpointConfig config, HandshakeRequest request, HandshakeResponse response)
+    @OnMessage
+    public String onMessage(String msg)
     {
-        HttpSession httpSession = (HttpSession)request.getHttpSession();
-        config.getUserProperties().put(HttpSession.class.getName(), httpSession);
+        // echo the message back to the remote
+        return msg;
     }
 }
