+++ jetty-websocket/javax-websocket-server/src/test/java/org/eclipse/jetty/websocket/javax/server/examples/StreamingEchoSocket.java
@@ -16,16 +16,17 @@
 //  ========================================================================
 //
 
-package examples;
+package org.eclipse.jetty.websocket.javax.server.examples;
 
-import org.eclipse.jetty.util.IO;
+import java.io.IOException;
+import java.io.Reader;
+import java.io.Writer;
 
 import javax.websocket.OnMessage;
 import javax.websocket.Session;
 import javax.websocket.server.ServerEndpoint;
-import java.io.IOException;
-import java.io.Reader;
-import java.io.Writer;
+
+import org.eclipse.jetty.util.IO;
 
 @ServerEndpoint("/echo")
 public class StreamingEchoSocket
